backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-whois>=0.8.0
xgboost>=2.0.0
